+++ src/components/RegisterView.js
@@ -1,17 +1,17 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import { loginThenGoToUserProfile as login } from "../actions";
+import { registerThenGoToUserProfile as register } from "../actions";
 import Spinner from "react-spinkit";
 import StickyHeader from "./StickyHeader";
 import { Form, FormField, Card, Button, Image } from "semantic-ui-react";
 import logo from "../squirrellogo.png";
 
-class LoginForm extends Component {
+class RegisterView extends Component {
   state = { username: "", password: "" };
 
-  handleLogin = e => {
+  handleRegister = e => {
     e.preventDefault();
-    this.props.login(this.state);
+    this.props.register(this.state);
   };
 
   handleChange = e => {
@@ -26,23 +26,42 @@ class LoginForm extends Component {
         <Card style={{ margin: "auto" }}>
           <Image src={logo} alt="" />
           <Card.Content>
-            <Form onSubmit={this.handleLogin}>
-              <h1 style={{ textAlign: "center" }}>Login</h1>
+            <Form onSubmit={this.handleRegister}>
+              <h1 style={{ textAlign: "center" }}>Register</h1>
               <FormField>
                 <Form.Input
                   placeholder="Username"
                   name="username"
+                  required
                   fluid
                   autoFocus
                   label="Username"
                   onChange={this.handleChange}
                 />
                 <Form.Input
+                  placeholder="New display name"
+                  name="displayName"
+                  required
+                  fluid
+                  label="Display name"
+                  onChange={this.handleChange}
+                />
+                <Form.Input
                   placeholder="Password"
                   name="password"
                   type="password"
+                  required
                   fluid
                   label="Password"
+                  onChange={this.handleChange}
+                />
+                <Form.Input
+                  placeholder="Confirm Password"
+                  name="password"
+                  type="password"
+                  required
+                  fluid
+                  label="Confirm Password"
                   onChange={this.handleChange}
                 />
                 {isLoading && <Spinner name="circle" color="blue" />}
@@ -68,7 +87,7 @@ class LoginForm extends Component {
 export default connect(
   ({ auth }) => ({
     isLoading: auth.loginLoading,
-    err: auth.loginError
+    err: auth.registerError
   }),
-  { login }
-)(LoginForm);
+  { register }
+)(RegisterView);
