+++ src/components/index.js
@@ -1,3 +1,6 @@
 export { default as App } from "./App";
-export { default as LoginForm } from "./LoginForm";
-export { default as UserProfile } from "./UserProfile";
+export { default as LoginView } from "./LoginView";
+export { default as ProfileView } from "./ProfileView";
+export { default as HomeView } from "./HomeView";
+export { default as RegisterView } from "./RegisterView";
+
